+++ systemtest/src/main/java/io/strimzi/systemtest/clients/KafkaClientProperties.java
@@ -192,15 +192,29 @@ class KafkaClientProperties {
                 .inNamespace(namespace)
                 .withName(externalBootstrapServiceName(clusterName))
                 .get();
+
         if (extBootstrapService == null) {
             throw new RuntimeException("Kafka cluster " + clusterName + " doesn't have an external bootstrap service");
         }
-        LoadBalancerIngress loadBalancerIngress = extBootstrapService.getStatus().getLoadBalancer().getIngress().get(0);
-        String result = loadBalancerIngress.getHostname();
-        if (result == null) {
-            result = loadBalancerIngress.getIp();
+
+        String extBootstrapServiceType = extBootstrapService.getSpec().getType();
+
+        if (extBootstrapServiceType.equals("NodePort")) {
+            int port = extBootstrapService.getSpec().getPorts().get(0).getNodePort();
+            String externalAddress = CLIENT.nodes().list().getItems().get(0).getStatus().getAddresses().get(0).getAddress();
+//            return extBootstrapService.getSpec().getClusterIP() + ":" + port;
+            return externalAddress + ":" + port;
+        } else if (extBootstrapServiceType.equals("LoadBalancer")) {
+            LoadBalancerIngress loadBalancerIngress = extBootstrapService.getStatus().getLoadBalancer().getIngress().get(0);
+            String result = loadBalancerIngress.getHostname();
+
+            if (result == null) {
+                result = loadBalancerIngress.getIp();
+            }
+            return result + ":9094";
+        } else {
+            throw new RuntimeException("Unexpected external bootstrap service for Kafka cluster " + clusterName);
         }
-        return result + ":9094";
     }
 
     /**
